Broadcast integration into main window
