+++ frontend/data/locale/en-US.ini
@@ -1561,6 +1561,18 @@ YouTube.Errors.invalidTransition="The attempted transition was invalid. This may
 YouTube.DocksRemoval.Title="Clear Legacy YouTube Browser Docks"
 YouTube.DocksRemoval.Text="These browser docks will be removed as deprecated:\n\n%1\nUse \"Docks/YouTube Live Control Room\" instead."
 
+# Restream Actions and Auth
+Restream.Actions.WindowTitle="Restream Broadcast Setup"
+Restream.Actions.BroadcastSelectTitle="Select Existing Broadcast"
+Restream.Actions.DashboardButton="Open Restream";
+Restream.Actions.BroadcastSelectButton="Select broadcast"
+Restream.Actions.BroadcastSelectAndStartButton="Select broadcast and start streaming"
+Restream.Actions.BroadcastScheduled="Scheduled"
+Restream.Actions.BroadcastLoadingTitle="Loading broadcast information..."
+Restream.Actions.BroadcastLoadingText="Loading broadcast information for %1, please wait..."
+Restream.Actions.BroadcastLoadingFailureTitle="Failed to load broadcast information"
+Restream.Actions.BroadcastLoadingFailureText="Failed to load broadcast information for %1\n\n%2: %3"
+
 # MultitrackVideo
 ConfigDownload.WarningMessageTitle="Warning"
 FailedToStartStream.MissingConfigURL="No config URL available for the current service"
