+++ frontend/widgets/OBSBasic_YouTube.cpp
@@ -129,87 +129,6 @@ void OBSBasic::ShowYouTubeAutoStartWarning()
 }
 #endif
 
-void OBSBasic::BroadcastButtonClicked()
-{
-	if (!broadcastReady || (!broadcastActive && !outputHandler->StreamingActive())) {
-		SetupBroadcast();
-		return;
-	}
-
-	if (!autoStartBroadcast) {
-#ifdef YOUTUBE_ENABLED
-		std::shared_ptr<YoutubeApiWrappers> ytAuth = dynamic_pointer_cast<YoutubeApiWrappers>(auth);
-		if (ytAuth.get()) {
-			if (!ytAuth->StartLatestBroadcast()) {
-				auto last_error = ytAuth->GetLastError();
-				if (last_error.isEmpty())
-					last_error = QTStr("YouTube.Actions.Error.YouTubeApi");
-				if (!ytAuth->GetTranslatedError(last_error))
-					last_error = QTStr("YouTube.Actions.Error.BroadcastTransitionFailed")
-							     .arg(last_error, ytAuth->GetBroadcastId());
-
-				OBSMessageBox::warning(this, QTStr("Output.BroadcastStartFailed"), last_error, true);
-				return;
-			}
-		}
-#endif
-		broadcastActive = true;
-		autoStartBroadcast = true; // and clear the flag
-
-		emit BroadcastStreamStarted(autoStopBroadcast);
-	} else if (!autoStopBroadcast) {
-#ifdef YOUTUBE_ENABLED
-		bool confirm = config_get_bool(App()->GetUserConfig(), "BasicWindow", "WarnBeforeStoppingStream");
-		if (confirm && isVisible()) {
-			QMessageBox::StandardButton button = OBSMessageBox::question(
-				this, QTStr("ConfirmStop.Title"), QTStr("YouTube.Actions.AutoStopStreamingWarning"),
-				QMessageBox::Yes | QMessageBox::No, QMessageBox::No);
-
-			if (button == QMessageBox::No)
-				return;
-		}
-
-		std::shared_ptr<YoutubeApiWrappers> ytAuth = dynamic_pointer_cast<YoutubeApiWrappers>(auth);
-		if (ytAuth.get()) {
-			if (!ytAuth->StopLatestBroadcast()) {
-				auto last_error = ytAuth->GetLastError();
-				if (last_error.isEmpty())
-					last_error = QTStr("YouTube.Actions.Error.YouTubeApi");
-				if (!ytAuth->GetTranslatedError(last_error))
-					last_error = QTStr("YouTube.Actions.Error.BroadcastTransitionFailed")
-							     .arg(last_error, ytAuth->GetBroadcastId());
-
-				OBSMessageBox::warning(this, QTStr("Output.BroadcastStopFailed"), last_error, true);
-			}
-		}
-#endif
-		broadcastActive = false;
-		broadcastReady = false;
-
-		autoStopBroadcast = true;
-		QMetaObject::invokeMethod(this, "StopStreaming");
-		emit BroadcastStreamReady(broadcastReady);
-		SetBroadcastFlowEnabled(true);
-	}
-}
-
-void OBSBasic::SetBroadcastFlowEnabled(bool enabled)
-{
-	emit BroadcastFlowEnabled(enabled);
-}
-
-void OBSBasic::SetupBroadcast()
-{
-#ifdef YOUTUBE_ENABLED
-	Auth *const auth = GetAuth();
-	if (IsYouTubeService(auth->service())) {
-		OBSYoutubeActions dialog(this, auth, broadcastReady);
-		connect(&dialog, &OBSYoutubeActions::ok, this, &OBSBasic::YouTubeActionDialogOk);
-		dialog.exec();
-	}
-#endif
-}
-
 #ifdef YOUTUBE_ENABLED
 YouTubeAppDock *OBSBasic::GetYouTubeAppDock()
 {
